Continued Chapter 7, began refactor of test logic
