+++ cr-api/.env
@@ -1,2 +1,2 @@
-SQLX_OFFLINE=true
+#SQLX_OFFLINE=true
 DATABASE_URL="postgres://postgres:password@localhost:5432/newsletter"
