+++ cr-api/configuration/base.yaml
@@ -11,3 +11,4 @@ email_client:
   base_url: "localhost"
   sender_email: "test@gmail.com"
   authorization_token: "my-secret-token"
+  timeout_milliseconds: 10000
