+++ cr-api/src/bin/main.rs
@@ -27,10 +27,12 @@ async fn main() -> hyper::Result<()> {
         .email_client
         .sender()
         .expect("Invalid sender email address.");
+    let timeout = configuration.email_client.timeout();
     let email_client = EmailClient::new(
         configuration.email_client.base_url,
         sender_email,
         configuration.email_client.authorization_token,
+        timeout
     );
 
     // Construct an address and port, get the port from the configuration settings
