+++ cr-api/tests/api/main.rs
@@ -1,0 +1,6 @@
+// tests/api/main.rs
+
+mod helpers;
+mod health_check;
+mod subscriptions;
+
